+++ trio_parallel/__init__.py
@@ -2,9 +2,10 @@
 
 from ._impl import (
     run_sync,
-    cache_scope,
+    open_worker_context,
+    WorkerContext,
     WorkerType,
     current_default_worker_limiter,
-    default_shutdown_grace_period,
+    atexit_shutdown_grace_period,
 )
 from ._abc import BrokenWorkerError